Move store into folder, exclude from coverage reports.
