+++ spec/src/modules/tracker-humanity.js
@@ -1,7 +1,7 @@
 const dotenv = require('dotenv');
 const chai = require('chai');
 const chaiAsPromised = require('chai-as-promised');
-const store = require('../../../src/store');
+const store = require('../../../src/store/store');
 const trackerHumanity = require('../../../src/modules/tracker-humanity');
 const helpers = require('../../mocha.helpers');
