+++ spec/src/modules/tracker-requests.js
@@ -2,7 +2,7 @@
 const dotenv = require('dotenv');
 const chai = require('chai');
 const chaiAsPromised = require('chai-as-promised');
-const store = require('../../../src/store');
+const store = require('../../../src/store/store');
 const trackerRequests = require('../../../src/modules/tracker-requests');
 const helpers = require('../../mocha.helpers');
