+++ src/modules/tracker-humanity.js
@@ -1,4 +1,4 @@
-const store = require('../store');
+const store = require('../store/store');
 
 const humanEvents = [
   'scroll',
